Make tests/implemented.py a proper module
New way to check implemented cards:

    $ python -m fireplace.implemented
