+++ fireplace/implemented/__init__.py
@@ -1,0 +1,64 @@
+import re
+import string
+from fireplace.utils import get_script_definition
+from hearthstone.enums import CardSet
+
+
+SOLVED_KEYWORDS = [
+	"Windfury", "Charge", "Divine Shield", "Taunt", "Stealth",
+	r"Can't be targeted by spells or Hero Powers\.",
+	r"Can't attack\.",
+	"Destroy any minion damaged by this minion.",
+	r"Your Hero Power deals \d+ extra damage.",
+	r"Spell Damage \+\d+",
+	r"Overload: \(\d+\)",
+]
+
+DUMMY_CARDS = (
+	"PlaceholderCard",  # Placeholder Card
+	"CS2_022e",  # Polymorph
+	"EX1_246e",  # Hexxed
+	"EX1_345t",  # Shadow of Nothing
+	"GAME_006",  # NOOOOOOOOOOOO
+	"LOEA04_27",  # Animated Statue
+	"Mekka4e",  # Transformed
+	"NEW1_025e",  # Bolstered (Unused)
+	"TU4c_005",  # Hidden Gnome
+	"TU4c_007",  # Mukla's Big Brother
+
+	# Dynamic buffs set by their parent
+	"CS2_236e",  # Divine Spirit
+	"EX1_304e",  # Consume (Void Terror)
+	"LOE_030e",  # Hollow (Unused)
+	"NEW1_018e",  # Treasure Crazed (Bloodsail Raider)
+)
+
+
+def cleanup_description(description):
+	ret = description
+	ret = re.sub("<i>.+</i>", "", ret)
+	ret = re.sub("(<b>|</b>)", "", ret)
+	ret = re.sub("(" + "|".join(SOLVED_KEYWORDS) + ")", "", ret)
+	ret = re.sub("<[^>]*>", "", ret)
+	exclude_chars = string.punctuation + string.whitespace
+	ret = "".join([ch for ch in ret if ch not in exclude_chars])
+	return ret
+
+
+def is_implemented(card):
+	description = cleanup_description(card.description)
+
+	if not description:
+		# Minions without card text or with basic abilities are implemented
+		return True
+	elif card.card_set == CardSet.CREDITS:
+		return True
+
+	if id in DUMMY_CARDS:
+		return True
+
+	carddef = get_script_definition(card.id)
+	if carddef:
+		return True
+
+	return False
